examples/integrations/autogen/requirements.txt
magic-ai-sdk>=0.2.0
pyautogen>=0.3.0
python-dotenv>=1.0.0
